requirements.txt
fastapi[standard]
uvicorn
pytesseract
pillow
pdf2image